Add JSON deserialization to HomeController
In the HomeController, a new method 'HelloWorld' was added that accepts a JSON string as a parameter. This JSON string is deserialized into an object using the Newtonsoft.Json library, which was added into this file. The method further extracts a value from the deserialized object to display in the response.